make travel time reflect distance on map
distance = sqrt( xDiff^2 + yDiff^2 ) where x and y values come from the differences between player position and travel-to-location in terms of SDL positions of the window

+++ src/game.h
@@ -69,7 +69,7 @@ class Game
   void updateGSD(); // GameStatsDisplay
   void clearBuilding();
   void scavengeBuilding();
-  int computeTravelTime();
+  float computeTravelTime();
 
   SDL_Point getBuildingCoord( int id );
